More compilation fixes for Linux.
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@157 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/ref_gl/r_rsurf.c
@@ -53,7 +53,7 @@ float   wTmu5Array[MAX_BATCH_SURFS][2];
 float   wTmu6Array[MAX_BATCH_SURFS][2];
 float	SurfColorArray[MAX_BATCH_SURFS][4];
 
-WORD	indexArray[MAX_BATCH_SURFS*3];
+int		indexArray[MAX_BATCH_SURFS*3];
 
 vec3_t	nTexArray[MAX_BATCH_SURFS];
 vec3_t	tTexArray[MAX_BATCH_SURFS];
